completionProvider: Drop 4 character check
We currently only try to provide inline completion if
the suggestion widget is active, if at least 4 characters
match the popup.

The problem is that suggestion widget doesn't have a lot of smarts to
figure out what to suggest, it mainly just picks from an alphabetically
sorted list, so assuming its suggestion is what the user wants is
dubious behavior unless the user has somehow made some motions to
indicate the suggestion widget selection is useful.

But, actually, the suggestion widget can be influenced by the inline
completion hint, which comes from a smart LLM, so rather than influencing
the inline completion from the suggestion widget, we should really
be influencing the suggestion widget from the inline completion.

As a first step toward achieving that goal, this commit reverts the
4 character logic that was added in

commit 304056d4995b086d989676db23ece7a45406e618

A subsequent commit will work reversing the dependencies between
inline completions and the suggestion widget.