completionProvider: Order suggestion widget after inline completion
Right now the suggestion widget will show up before the inline
completion is processed. This causes a couple of problems:

1. It is not informed by the results of an LLM so its
   suggestions are often off base.

2. It affects the editor buffer so can mislead the LLM on what
   sort of inline completions to provide.

This commit flips things on their head: We now inhibit the
suggestion widget until the inline completion is generated and
even then, only uninhibit the the suggestion widget if it has
at least one suggestion that is compatible with the inline
suggestion.

The overarching idea is that the LLM is the "smart" output, so
it should be what drives completion, not the context-naïve
alphabetically ordered list of completions provides by
intellisense.