Granite.Code: disable tool autodetection for Granite 3
Granite 3 doesn't work very well with agent model; it tends
to tell the user to use the tool rather than actually using
it. Until the model is improved, or we spend the effort to
drive it better, don't detect it as supporting tools.

This can be overridden using capabilities: [tool_use] in
a model section in config.yaml.