core/util: Add class for managing code fragments
Managing and integrating chunks of source code into the
editor window is easy to get wrong, especially if there is
overlap between what's there and what's getting integrated.

To help facilitate that kind of work, this commit adds a new
SourceFragment class. It encapsulates a chunk of code and provides
ways to access it from the top and the bottom, and ways to split
it into an array of lines, optionally trimming whitespace and
blank lines in the process.

SourceFragment also comes with a getRemainingCompletion
method that determines if the bottom of the fragment matches
the top of the passed in fragment. It then returns what's not
yet integrated.

There is also a getAsTruncatedFragment function that will chop
off the specified tail of the fragment, optionally ignoring
whitespace.