autocomplete: Massage granite output in place more thoroughly
The granite models sometimes fail to completely follow instructions
in two ways:

1. They'll often repeat the bit of the prefix on the same line as
the completion.

2. They'll sometimes repeat the suffix and then even go further and
hallucinate code after the suffix.

The code already has a mitigation in place for 1., but nothing to
handle 2. Furthermore, the mitigation for 1. is open coded, since it
predates SourceFragement.

This commit changes the code over to use SourceFragment and at the
same time, addresses 2 by chopping off the spurious completion text
at the suffix start.