Fix: change the mechanism of checking incomoatible extensions
Since the build script will replac "Continue.continue" to "redhat.granitecode",
here we change the mechanism to grab the extension id during the activation and
remove the current extension id from the incompatible extensions list.
Therefore, we can mitigate two situations that we block ourselves:
  1. The built version (Due to replacing Continue.continue to
     redhat.granitecode)
  2. During dev mode (We are Continue while in dev mode)

We do need to change the build script to skip this file so it can work
properly.