Fix images in Granite Setup Wizard in dev mode
When the Granite Setup Wizard is run in developer mode (from
the Extension Host by the Play button in VS Code), assets
are hosted by the VITE webserver instead of directly accessed on
the filesystem.

There's already various places in the code to handle this asset
resolution differently for developer mode versus release mode, but
image assets in the Setup Wizard missed that special handling.

This commit adds logic to tell the Setup Wizard to load from the
web server in developer mode, and also adds code to update the
content security policy to allow getting images in that manner.