Feat: include .graniteignore and .graniterules for Granite.Code
This feature add the functionality for Granite.Code to check .graniteignore files for ignore files,
and .graniterules files for customized system prompts (note: .continuerules will be deprecated in a
future release of Continue).

Note: The ignore files affects what files will be excluded for indexing codebase. They also affect
some file searching functions.

Here is the detail for this new mechanism:
	1. For .graniteignore files, if it exists at current folder, it will be the only ignore
           file we account for. The .continueignore file at the same directory will be ignored.
	2. The above rule is also applied to .graniterules. The only difference is .continuerules and
           .graniterules will only be detected at the workspace directory.

I also added test cases for the shouldIgnore and walkDir functions to ensure they behave as expected
and correctly handle file exclusion based on ignore rules in .graniteignore.

I put a "test search workspace!" button at the bottom right corner to trigger the search files
function and send the result to the modal dialog so this feature can be easily tested. I'll remove
this button when we're ready to merge.

https://github.com/Granite-Code/granite-code/issues/129